requirements.txt
flask
pymongo
flask_restx
boto3
werkzeug
awscli
Werkzeug
PyJWT
flask_migrate
celery
mongoengine
flask_mongoengine
